aperture/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
supabase==2.3.0
asyncpg==0.29.0

# LLM Providers
openai==1.10.0
anthropic==0.18.1
litellm==1.24.3

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tiktoken==0.6.0
nanoid==2.0.0

# Web UI
jinja2==3.1.3

# Background jobs (future)
# celery==5.3.6
# redis==5.0.1
